Created four React components to house the main functions of the page
