+++ src/Components/Header.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Header = (props) => {
+  return (
+    <header className="App-header">
+        Read Around the World Book Search
+    </header>
+  );
+}
+
+export default Header;
