--- src/Main.jsx
@@ -1,18 +1,0 @@
-import React, { Component } from 'react';
-
-
-export default class Main extends Component {
-  constructor(props) {
-    super(props)
-  }
-
-  this.state = {}
-
-  render () {
-    return (
-      <div className="App-body">
-        Herro, I am Main!!
-      </div>
-    );
-  }
-}
